Update dependencies, eliminate broken webpack plugin, begin browser refactorings
