+++ package.json
@@ -20,22 +20,23 @@
         "check:lockindent": "node ./build/checkLockIndent.js"
     },
     "devDependencies": {
-        "@types/glob": "^7.1.3",
-        "@types/node": "^12.20.15",
-        "@types/yargs": "^16.0.3",
-        "@typescript-eslint/eslint-plugin": "^4.26.1",
-        "@typescript-eslint/parser": "^4.26.1",
+        "@types/glob": "^7.1.4",
+        "@types/node": "^12.20.16",
+        "@types/yargs": "^16.0.4",
+        "@typescript-eslint/eslint-plugin": "^4.28.3",
+        "@typescript-eslint/parser": "^4.28.3",
         "detect-indent": "^6.1.0",
-        "eslint": "^7.28.0",
+        "eslint": "^7.30.0",
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-simple-import-sort": "^7.0.0",
         "glob": "^7.1.7",
+        "jsonc-parser": "^3.0.0",
         "lerna": "^4.0.0",
-        "npm-check-updates": "^11.6.0",
+        "npm-check-updates": "^11.8.3",
         "p-queue": "^6.6.2",
-        "prettier": "2.3.1",
+        "prettier": "2.3.2",
         "syncpack": "^5.7.11",
-        "typescript": "~4.3.2",
+        "typescript": "~4.3.5",
         "yargs": "^16.2.0"
     }
 }
