+++ packages/pyright-internal/package.json
@@ -16,12 +16,12 @@
     },
     "dependencies": {
         "@iarna/toml": "2.2.5",
-        "@yarnpkg/fslib": "2.5.0-rc.5",
-        "@yarnpkg/libzip": "2.2.2-rc.3",
+        "@yarnpkg/fslib": "2.5.0-rc.9",
+        "@yarnpkg/libzip": "2.2.2-rc.7",
         "assert": "^2.0.0",
         "chalk": "^4.1.1",
-        "chokidar": "^3.5.1",
-        "command-line-args": "^5.1.1",
+        "chokidar": "^3.5.2",
+        "command-line-args": "^5.1.3",
         "jsonc-parser": "^3.0.0",
         "leven": "^3.1.0",
         "source-map-support": "^0.5.19",
@@ -34,14 +34,14 @@
         "vscode-uri": "^3.0.2"
     },
     "devDependencies": {
-        "@types/command-line-args": "^5.0.0",
-        "@types/jest": "^26.0.23",
-        "@types/node": "^12.20.15",
-        "@types/tmp": "^0.2.0",
-        "jest": "^27.0.4",
+        "@types/command-line-args": "^5.0.1",
+        "@types/jest": "^26.0.24",
+        "@types/node": "^12.20.16",
+        "@types/tmp": "^0.2.1",
+        "jest": "^27.0.6",
         "jest-junit": "^12.2.0",
         "shx": "^0.3.3",
         "ts-jest": "^27.0.3",
-        "typescript": "~4.3.2"
+        "typescript": "~4.3.5"
     }
 }
