+++ packages/pyright/package.json
@@ -23,14 +23,13 @@
         "webpack": "webpack --mode development --progress"
     },
     "devDependencies": {
-        "@types/copy-webpack-plugin": "^8.0.0",
-        "@types/node": "^12.20.15",
-        "copy-webpack-plugin": "^9.0.0",
+        "@types/copy-webpack-plugin": "^8.0.1",
+        "@types/node": "^12.20.16",
+        "copy-webpack-plugin": "^9.0.1",
         "shx": "^0.3.3",
         "ts-loader": "^9.2.3",
-        "tsconfig-paths-webpack-plugin": "^3.5.1",
-        "typescript": "~4.3.2",
-        "webpack": "^5.38.1",
+        "typescript": "~4.3.5",
+        "webpack": "^5.44.0",
         "webpack-cli": "^4.7.2"
     },
     "files": [
