+++ packages/vscode-pyright/package.json
@@ -774,17 +774,16 @@
         "vscode-languageserver-protocol": "3.16.0"
     },
     "devDependencies": {
-        "@types/copy-webpack-plugin": "^8.0.0",
-        "@types/node": "^12.20.15",
+        "@types/copy-webpack-plugin": "^8.0.1",
+        "@types/node": "^12.20.16",
         "@types/vscode": "~1.54.0",
-        "copy-webpack-plugin": "^9.0.0",
+        "copy-webpack-plugin": "^9.0.1",
         "detect-indent": "^6.1.0",
         "shx": "^0.3.3",
         "ts-loader": "^9.2.3",
-        "tsconfig-paths-webpack-plugin": "^3.5.1",
-        "typescript": "~4.3.2",
-        "vsce": "^1.93.0",
-        "webpack": "^5.38.1",
+        "typescript": "~4.3.5",
+        "vsce": "^1.95.1",
+        "webpack": "^5.44.0",
         "webpack-cli": "^4.7.2"
     }
 }
